refactor(legacy): Define the model for legacy support
