+++ src/legacy_model/mod.rs
@@ -1,0 +1,7 @@
+mod config;
+mod document;
+mod frontmatter;
+
+pub use self::config::GlobalConfig;
+pub use self::document::DocumentBuilder;
+pub use self::frontmatter::FrontmatterBuilder;
