+++ src/lib.rs
@@ -101,7 +101,7 @@ mod new;
 mod template;
 
 pub mod jekyll;
-pub mod legacy;
+pub mod legacy_model;
 mod syntax_highlight;
 
 pub use syntax_highlight::{list_syntax_themes, list_syntaxes};
